pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcriber"
version = "1.0.0"
description = "Audio transcription with speaker diarization"
authors = [{name = "Transcription Tool"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "openai-whisper",
    "pyannote-audio",
    "python-dotenv",
]

[project.scripts]
transcribe = "transcriber.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"